requirements.txt
# Satya Vaani - Requirements
# Install: pip install -r requirements.txt

# Core ML
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.6.0
accelerate>=0.28.0

# Vector Search
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU systems

# NLP
spacy>=3.7.0
stanza>=1.7.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.33.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0

# Indic NLP (optional, for deeper Indic language support)
# indic-nlp-library>=0.92

# Testing
pytest>=8.0.0
